whatsapp_metro_bot/requirements.txt
fastapi
uvicorn
requests
qrcode
pillow
python-dotenv
rasa
twilio
gunicorn
